moved spotify info over to ArtistDetails.js, artist spotify search currently not working
