+++ venu/src/components/stylesheets/ArtistDetails.css
@@ -33,3 +33,27 @@
   font-size: 16px;
   margin-top: 5px;
 }
+
+{/*Spotify stuff starts here*/}
+
+.search-tracks-button {
+  font-size: 16px;
+  font-weight: bold;
+  padding-top: 10px;
+}
+
+.login-link {
+  font-size: x-small;
+}
+
+.flex-container {
+  background-color: #ffff;
+  padding: 25px;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 20px;
+  margin-bottom: 5px;
+}
