+++ venu/src/components/stylesheets/EventDetails.css
@@ -34,28 +34,6 @@
   margin-top: 5px;
 }
 
-.search-tracks-button {
-  font-size: 16px;
-  font-weight: bold;
-  padding-top: 10px;
-}
-
-.login-link {
-  font-size: x-small;
-}
-
-.flex-container {
-  background-color: #ffff;
-  padding: 25px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 20px;
-  margin-bottom: 5px;
-}
-
 .artist-details-message {
   margin-left: 8px;
   font-size: 12px;
